Templatized all classes in pointer.cpp.
With templates, we don't have to pass the size of nodes, we can just
declare that the memory layer is for type <Node> or <int> or any other
defined type.  This makes it more extensible and also a lot easier to
use since we can just store an array with flat indexing (first node will
be at index 0, second will be at index 1, etc.), and we don't have to do
any clever indexing and offsetting.